chore: rename files, add docs
