+++ src/index.ts
@@ -3,8 +3,8 @@ import path from "node:path";
 import dotenv from "dotenv";
 import fs from "node:fs";
 import { fileURLToPath, pathToFileURL } from "node:url";
-import { DiscordClient } from "./discordClient.mjs";
-import type { Command, Event } from "./types.mjs";
+import { DiscordClient } from "./discord-client.js";
+import type { Command, Event } from "./types.js";
 
 dotenv.config();
 // Some hack to get __dirname to work in modules
@@ -16,23 +16,18 @@ const client = new DiscordClient({ intents: [GatewayIntentBits.Guilds] });
 {
     // Iterate over every command/folder
     const commandsPath = path.join(__dirname, "commands");
+    const command_sources = ["./commands/countdown", "./commands/ping"] as const;
+    const val = (await import("./commands/countdown")).default;
+
+    const commands = Promise.all(command_sources.map(source => import(source)));
     const commandFiles = fs.readdirSync(commandsPath).filter(
         file => file.endsWith(".mjs"), // Compiled to js!
     );
     // Iterate over file in said folder
-    for (const file of commandFiles) {
-        // Import contents
-        const filePath = path.join(commandsPath, file);
-        const resolvedPath = pathToFileURL(filePath).href;
-        const command: Command = (await import(resolvedPath)).default;
+    for (const file of command_sources) {
+        const command = (await import(file)).default;
         // Set a new item in the Collection with the key as the command name and the value as the exported module
-        if (command && `data` in command && `execute` in command) {
-            // Bind command
-            client.commands.set(command.data.name, command);
-        } else {
-            // Invalid/missing contents
-            console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
-        }
+        client.commands.set(command.data.name, command);
     }
 }
