+++ tsconfig.json
@@ -1,12 +1,19 @@
 {
-  "compilerOptions": {
-    "module": "esnext",
-    "esModuleInterop": true,
-    "target": "es2017",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "lib": ["es2017"]
+    "compilerOptions": {
+        "module": "esnext",
+        "target": "es2017",
+        "moduleResolution": "node",
+        "esModuleInterop": true,
+        "sourceMap": true,
+        "strict": true,
+        "forceConsistentCasingInFileNames": true,
+        "noImplicitAny": true,
+        "outDir": "dist",
+        "rootDir": "./src",
+        "paths": {
+            "@/*": ["./src/*"]
+        },
+    },
+    "include": ["src/**/*"],
+    "lib": ["es2017"]
 }
